Move to `map_cmp.go` and revoke tinygo
